Refactor: Extract streaming and non-streaming handlers to separate files
- Create src/handlers/streamingHandler.js for streaming response logic
- Create src/handlers/nonStreamingHandler.js for non-streaming response logic
- Create src/utils/openaiResponse.js for OpenAI response formatting
- Reduce chatCompletions.js from 215 to 143 lines (~33% reduction)
- Improve separation of concerns and testability
- All tests pass (208 tests, 93.79% coverage)
